Created the user login service with JWT and created the car by the user.

+++ src/main/java/com/arunpragash/car_rental/repository/CarRepository.java
@@ -1,8 +1,11 @@
 package com.arunpragash.car_rental.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.arunpragash.car_rental.model.table.Car;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
+   
 }
